+++ FertiWeb/src/main/java/persistencia/repositorio/AplicacionRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import persistencia.entidad.AplicacionEntidad;
 
-public interface AplicacionRepository extends JpaRepository<AplicacionEntidad, Long> {
+public interface AplicacionRepository extends JpaRepository<AplicacionEntidad, Integer> {
 
 }
