+++ FertiWeb/src/main/java/persistencia/repositorio/CultivoRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import persistencia.entidad.CultivoEntidad;
 
-public interface CultivoRepository extends JpaRepository<CultivoEntidad, Long> {
+public interface CultivoRepository extends JpaRepository<CultivoEntidad, Integer> {
 
 }
