+++ FertiWeb/src/main/java/persistencia/repositorio/LoteRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import persistencia.entidad.LoteEntidad;
 
-public interface LoteRepository extends JpaRepository<LoteEntidad, Long> {
+public interface LoteRepository extends JpaRepository<LoteEntidad, Integer> {
 
 }
