+++ FertiWeb/src/main/java/persistencia/repositorio/ResultadoRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import persistencia.entidad.ResultadoEntidad;
 
-public interface ResultadoRepository extends JpaRepository<ResultadoEntidad, Long> {
+public interface ResultadoRepository extends JpaRepository<ResultadoEntidad, Integer> {
 
 }
