+++ FertiWeb/src/main/java/persistencia/repositorio/TipoUsuarioRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import persistencia.entidad.TipoUsuarioEntidad;
 
-public interface TipoUsuarioRepository extends JpaRepository<TipoUsuarioEntidad, Long> {
+public interface TipoUsuarioRepository extends JpaRepository<TipoUsuarioEntidad, Integer> {
 
 }
